requirements.txt
# ═══════════════════════════════════════════════════════════════════════════
# File    : requirements.txt
# Desc    : Production dependencies — versi pin eksak sesuai Blueprint Appendix B.
# Step    : STEP 3 — Backend Setup
# Ref     : Blueprint Appendix B
# ═══════════════════════════════════════════════════════════════════════════

fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
sqlalchemy==2.0.30
alembic==1.13.1
asyncpg==0.29.0
celery==5.4.0
redis==5.0.4
httpx==0.27.0
openai==1.30.1
anthropic>=0.40.0
google-generativeai>=0.8.0
python-jose[cryptography]==3.3.0
structlog==24.1.0
sentry-sdk[fastapi]==2.3.1
python-docx==1.1.2
weasyprint==62.3
cairosvg==2.7.1
midtransclient==1.4.2
resend==2.0.0
